+++ src/App.tsx
@@ -1,106 +1,25 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
-import QuestionContainer from "./QuestionContainer";
-import Question from "./Question";
-import Email from "./Email";
-import Perspective from "./Perspective";
+import QuestionsPage from "./QuestionsPage";
+import ResultPage from "./ResultPage";
 
 import "./index.scss";
 
 const App = () => {
-  const [welcomeMessage, setWelcomeMessage] = useState<string>("");
-  const [data, setData] = useState<any>({});
-
-  useEffect(() => {
-    fetch("/hello").then(async (response) => {
-      try {
-        const data = await response.json();
-        setWelcomeMessage(data.message);
-      } catch {
-        setWelcomeMessage("");
-      }
-    });
-  }, []);
-
-  const questions = [
-    "You find it takes effort to introduce yourself to other people",
-    "You consider yourself more practical than creative.",
-    "Winning a debate matters less to you than making sure no one gets upset.",
-    "You get energized going to social events that involve many interactions.",
-    "You often spend time exploring unrealistic and impractical yet intriguing ideas.",
-    "Deadlines seem to you to be of relative rather than absolute importance.",
-    "Logic is usually more important than heart when it comes to making important decisions.",
-    "Your home and work environments are quite tidy.",
-    "You do not mind being at the center of attention.",
-    "Keeping your options open is more important than having a to-do list.",
-  ];
-
-  const perspectives = [
-    { name: "EI", left: "Introversion (I)", right: "Extroversion (E)" },
-    { name: "SN", left: "Sensing (S)", right: "Intuition (N)" },
-    { name: "TF", left: "Thinking (T)", right: "Feeling (F)" },
-    { name: "JP", left: "Judging (J)", right: "Perceiving (P)" },
-  ];
-
-  const handleDataUpdate = (key: string, value: string | number) => {
-    const newData = { ...data, [key]: value };
-    setData(newData);
-  };
-
-  const handleEmailUpdate = (value: string) => {
-    handleDataUpdate("email", value);
-  };
-
-  const handleSubmit = () => {
-    console.log("Current data", data);
-  };
-
   return (
-    <div className="App">
-      {/* <div>
-        <div className="discover-title">
-          <p>Discover Your Perspective</p>
-          <span>
-            Complete the 7 min test and get a detailed report of your lenses on
-            the world.
-          </span>
-        </div>
-        <div className="questions-container">
-          <div className="questions">
-            {questions.map((question, index) => (
-              <QuestionContainer key={index + 1}>
-                <Question
-                  questionKey={index.toString()}
-                  updateValue={handleDataUpdate}
-                  question={question}
-                />
-              </QuestionContainer>
-            ))}
-            <QuestionContainer>
-              <Email updateEmail={handleEmailUpdate} />
-            </QuestionContainer>
-          </div>
-          <button className="save-button" onClick={handleSubmit}>
-            {"Save & Continue"}
-          </button>
-        </div>
-      </div> */}
-      <div className="result-page">
-        <div className="result-title">
-          <div>Your Perspective</div>
-          <span>Your Perspective type is ENTI</span>
-        </div>
-        <div>
-          {perspectives.map((perspective) => (
-            <Perspective
-              left={perspective.left}
-              right={perspective.right}
-              result={1}
-            />
-          ))}
-        </div>
+    <Router>
+      <div className="App">
+        <Switch>
+          <Route path="/result">
+            <ResultPage />
+          </Route>
+          <Route path="/">
+            <QuestionsPage />
+          </Route>
+        </Switch>
       </div>
-    </div>
+    </Router>
   );
 };
