+++ src/index.scss
@@ -31,10 +31,11 @@ body {
   display: flex;
   flex-direction: column;
   align-items: center;
+  padding-bottom: 3rem;
 
   .questions {
     width: 740px;
-    margin-bottom: 72px;
+    margin-bottom: 3rem;
   }
 }
 
@@ -122,6 +123,10 @@ body {
       font-size: 16px;
       color: #343a40;
     }
+  }
+
+  .perspectives {
+    width: 40%;
   }
 }
